Move TestLoadingSeriesChunkRefsSetIterator
